+++ Checkers/Checkers/Board.h
@@ -12,7 +12,7 @@ public:
 	// Размер доски.
 	const int BoardSize;
 
-	CBoard( size_t _BoardSize = 10, size_t _startNumberOfCheckers = 20 );
+	CBoard( int _BoardSize = 10, int _startNumberOfCheckers = 20 );
 
 	std::vector<CField>& GetBoard() { return playBoard; };
 
@@ -22,10 +22,10 @@ public:
 private:
 	
 	// Количество шашек у каждого игрока в начальный момент игры.
-	const size_t startNumberOfCheckers;
-
+	const int startNumberOfCheckers;
 	// Описывает поле для игры, содержит описания только черных, которые нумеруются в соответствии с определенными правилами.
 	std::vector<CField> playBoard;
 
+	// Создание доски для новой игры.
 	void generateNewBoard();
 };
